Implemente the depth buffer.
Now only the pixels whose depth value smaller than that in the same place
in depth buffer will be drawed.

Signature: Bruce Gan
Date: Feb 22

+++ Software_Render/include/framebuffer.h
@@ -6,5 +6,6 @@ public:
 	int height;
 	int bpp;
 	unsigned char* pixels;
-	unsigned short* depth;
+	float* depth;
 };
+
